Fix compile warnings in checkers and generators

+++ checkers/caseicmp.cpp
@@ -74,9 +74,9 @@ int main(int argc, char* argv[])
             quitf(_wa, "Expected %s found %s [test case %d]", vtos(ja[i]).c_str(), vtos(pa[i]).c_str(), i + 1);
 
     if (ja.size() != pa.size())
-        quitf(_pe, "Expected %d test case(s) but found %d", ja.size(), pa.size());
+        quitf(_pe, "Expected %u test case(s) but found %u", (unsigned int)(ja.size()), (unsigned int)(pa.size()));
 
-    string message = format("%d case(s):", ja.size());
+    string message = format("%u case(s):", (unsigned int)(ja.size()));
     if (ja.size() <= 5)
     {
         forn(i, ja.size())
